chore: move OkHttp3 interceptors into remote directory
